+++ api/responses/response_wrapper.py
@@ -3,3 +3,8 @@ def paginated_response(data, total_count: int):
         "data": data,
         "total_count": total_count,
     }
+
+def error_response(status, error):
+    return {
+        "error": error,
+    }, status