+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Added
 - Added pagination limit and offset to base repo get_all method
 - Added `get_all` route for `tyre_model` and response objects
+- Added `get_by_id` route for `tyre_model`
 ### Changed
 - Fixed `.env.dist` to match required environment variables
 - Omitted `database/session.py` and `database/extensions.py` from test coverage report